+++ packages/hardhat-core/src/internal/core/jsonrpc/types/base-types.ts
@@ -1,17 +1,21 @@
-import { BN, bufferToHex, isValidAddress, toBuffer } from "ethereumjs-util";
+import {
+  bufferToHex,
+  isValidAddress,
+  toBuffer,
+} from "@nomicfoundation/ethereumjs-util";
 import * as t from "io-ts";
 
+import * as BigIntUtils from "../../../util/bigint";
 import { assertHardhatInvariant, HardhatError } from "../../errors";
 import { ERRORS } from "../../errors-list";
 
 const ADDRESS_LENGTH_BYTES = 20;
 const HASH_LENGTH_BYTES = 32;
 
-export const rpcQuantity = new t.Type<BN>(
+export const rpcQuantity = new t.Type<bigint>(
   "QUANTITY",
-  BN.isBN,
-  (u, c) =>
-    isRpcQuantityString(u) ? t.success(new BN(toBuffer(u))) : t.failure(u, c),
+  BigIntUtils.isBigInt,
+  (u, c) => (isRpcQuantityString(u) ? t.success(BigInt(u)) : t.failure(u, c)),
   t.identity
 );
 
@@ -29,14 +33,14 @@ export const rpcHash = new t.Type<Buffer>(
   t.identity
 );
 
-export const rpcStorageSlot = new t.Type<BN>(
+export const rpcStorageSlot = new t.Type<bigint>(
   "Storage slot",
-  BN.isBN,
+  BigIntUtils.isBigInt,
   validateStorageSlot,
   t.identity
 );
 
-function validateStorageSlot(u: unknown, c: t.Context): t.Validation<BN> {
+function validateStorageSlot(u: unknown, c: t.Context): t.Validation<bigint> {
   if (typeof u !== "string") {
     return t.failure(
       u,
@@ -61,7 +65,7 @@ function validateStorageSlot(u: unknown, c: t.Context): t.Validation<BN> {
     );
   }
 
-  return t.success(new BN(toBuffer(u)));
+  return t.success(BigInt(u));
 }
 
 export const rpcAddress = new t.Type<Buffer>(
@@ -78,10 +82,10 @@ export const rpcUnsignedInteger = new t.Type<number>(
   t.identity
 );
 
-export const rpcQuantityAsNumber = new t.Type<BN>(
+export const rpcQuantityAsNumber = new t.Type<bigint>(
   "Integer",
-  BN.isBN,
-  (u, c) => (isInteger(u) ? t.success(new BN(u)) : t.failure(u, c)),
+  BigIntUtils.isBigInt,
+  (u, c) => (isInteger(u) ? t.success(BigInt(u)) : t.failure(u, c)),
   t.identity
 );
 
@@ -99,10 +103,10 @@ export const rpcFloat = new t.Type<number>(
  * fits in a number.
  */
 export function rpcQuantityToNumber(quantity: string): number {
-  return rpcQuantityToBN(quantity).toNumber();
+  return Number(rpcQuantityToBigInt(quantity));
 }
 
-export function rpcQuantityToBN(quantity: string): BN {
+export function rpcQuantityToBigInt(quantity: string): bigint {
   // We validate it in case a value gets here through a cast or any
   if (!isRpcQuantityString(quantity)) {
     throw new HardhatError(ERRORS.NETWORK.INVALID_RPC_QUANTITY_VALUE, {
@@ -110,26 +114,25 @@ export function rpcQuantityToBN(quantity: string): BN {
     });
   }
 
-  const buffer = toBuffer(quantity);
-  return new BN(buffer);
+  return BigInt(quantity);
 }
 
-export function numberToRpcQuantity(n: number | BN): string {
+export function numberToRpcQuantity(n: number | bigint): string {
   assertHardhatInvariant(
-    typeof n === "number" || BN.isBN(n),
+    typeof n === "number" || typeof n === "bigint",
     "Expected number"
   );
 
   return `0x${n.toString(16)}`;
 }
 
-export function numberToRpcStorageSlot(n: number | BN): string {
+export function numberToRpcStorageSlot(n: number | bigint): string {
   assertHardhatInvariant(
-    typeof n === "number" || BN.isBN(n),
+    typeof n === "number" || typeof n === "bigint",
     "Expected number"
   );
 
-  return `0x${n.toString(16).padStart(64, "0")}`;
+  return `0x${BigIntUtils.toEvmWord(n)}`;
 }
 
 /**
@@ -137,11 +140,11 @@ export function numberToRpcStorageSlot(n: number | BN): string {
  * represents a value fits in a number.
  */
 export function rpcDataToNumber(data: string): number {
-  return rpcDataToBN(data).toNumber();
+  return Number(rpcDataToBigInt(data));
 }
 
-export function rpcDataToBN(data: string): BN {
-  return new BN(rpcDataToBuffer(data));
+export function rpcDataToBigInt(data: string): bigint {
+  return data === "0x" ? 0n : BigInt(data);
 }
 
 export function bufferToRpcData(