+++ packages/hardhat-core/src/internal/hardhat-network/stack-traces/solidityTracer.ts
@@ -1,4 +1,4 @@
-import { ERROR } from "@ethereumjs/vm/dist/exceptions";
+import { ERROR } from "@nomicfoundation/ethereumjs-evm/dist/exceptions";
 import { ReturnData } from "../provider/return-data";
 
 import {
@@ -31,9 +31,6 @@ import {
   SolidityStackTraceEntry,
   StackTraceEntryType,
 } from "./solidity-stack-trace";
-
-export const SUPPORTED_SOLIDITY_VERSION_RANGE = "<=0.8.16";
-export const FIRST_SOLC_VERSION_SUPPORTED = "0.5.1";
 
 export class SolidityTracer {
   private _errorInferrer = new ErrorInferrer();