+++ packages/hardhat-core/src/internal/solidity/compiler/index.ts
@@ -1,12 +1,19 @@
-import { exec } from "child_process";
+import { execFile } from "child_process";
 import * as fs from "fs";
+import { CompilerInput, CompilerOutput } from "../../../types";
+import { HardhatError } from "../../core/errors";
+import { ERRORS } from "../../core/errors-list";
 
-export class Compiler {
+export interface ICompiler {
+  compile(input: CompilerInput): Promise<CompilerOutput>;
+}
+
+export class Compiler implements ICompiler {
   private _loadedSolc?: any;
 
   constructor(private _pathToSolcJs: string) {}
 
-  public async compile(input: any) {
+  public async compile(input: CompilerInput) {
     const solc = await this.getSolc();
 
     const jsonOutput = solc.compile(JSON.stringify(input));
@@ -52,26 +59,31 @@ export class Compiler {
   }
 }
 
-export class NativeCompiler {
+export class NativeCompiler implements ICompiler {
   constructor(private _pathToSolc: string) {}
 
-  public async compile(input: any) {
+  public async compile(input: CompilerInput) {
     const output: string = await new Promise((resolve, reject) => {
-      const process = exec(
-        `${this._pathToSolc} --standard-json`,
-        {
-          maxBuffer: 1024 * 1024 * 500,
-        },
-        (err, stdout) => {
-          if (err !== null) {
-            return reject(err);
+      try {
+        const process = execFile(
+          this._pathToSolc,
+          [`--standard-json`],
+          {
+            maxBuffer: 1024 * 1024 * 500,
+          },
+          (err, stdout) => {
+            if (err !== null) {
+              return reject(err);
+            }
+            resolve(stdout);
           }
-          resolve(stdout);
-        }
-      );
+        );
 
-      process.stdin!.write(JSON.stringify(input));
-      process.stdin!.end();
+        process.stdin!.write(JSON.stringify(input));
+        process.stdin!.end();
+      } catch (e: any) {
+        throw new HardhatError(ERRORS.SOLC.CANT_RUN_NATIVE_COMPILER, {}, e);
+      }
     });
 
     return JSON.parse(output);