+++ packages/hardhat-core/test/internal/hardhat-network/provider/modules/eth/methods/call.ts
@@ -1,13 +1,12 @@
 import { assert } from "chai";
-import { BN } from "ethereumjs-util";
 import { Client } from "undici";
 
 import {
   numberToRpcQuantity,
   rpcDataToNumber,
   rpcQuantityToNumber,
-  rpcDataToBN,
-  rpcQuantityToBN,
+  rpcDataToBigInt,
+  rpcQuantityToBigInt,
 } from "../../../../../../../src/internal/core/jsonrpc/types/base-types";
 import { getCurrentTimestamp } from "../../../../../../../src/internal/hardhat-network/provider/utils/getCurrentTimestamp";
 import { workaroundWindowsCiFailures } from "../../../../../../utils/workaround-windows-ci-failures";
@@ -515,7 +514,7 @@ contract C {
 
           const CALLER = DEFAULT_ACCOUNTS_ADDRESSES[2];
           let contractAddress: string;
-          let ethBalance: BN;
+          let ethBalance: bigint;
 
           function deployContractAndGetEthBalance() {
             beforeEach(async function () {
@@ -524,7 +523,7 @@ contract C {
                 deploymentBytecode
               );
 
-              ethBalance = rpcQuantityToBN(
+              ethBalance = rpcQuantityToBigInt(
                 await this.provider.send("eth_getBalance", [CALLER])
               );
               assert.notEqual(ethBalance.toString(), "0");
@@ -546,14 +545,14 @@ contract C {
               ]);
 
               assert.equal(
-                rpcDataToBN(balanceResult).toString(),
+                rpcDataToBigInt(balanceResult).toString(),
                 ethBalance.toString()
               );
             });
 
             it("Should use any provided gasPrice", async function () {
-              const gasLimit = 200_000;
-              const gasPrice = 2;
+              const gasLimit = 200_000n;
+              const gasPrice = 2n;
 
               const balanceResult = await this.provider.send("eth_call", [
                 {
@@ -565,10 +564,9 @@ contract C {
                 },
               ]);
 
-              assert.isTrue(
-                rpcDataToBN(balanceResult).eq(
-                  ethBalance.subn(gasLimit * gasPrice)
-                )
+              assert.equal(
+                rpcDataToBigInt(balanceResult),
+                ethBalance - gasLimit * gasPrice
               );
             });
           });
@@ -635,7 +633,7 @@ contract C {
                 ]);
 
                 assert.equal(
-                  rpcDataToBN(balanceResult).toString(),
+                  rpcDataToBigInt(balanceResult).toString(),
                   ethBalance.toString()
                 );
               });
@@ -653,7 +651,7 @@ contract C {
                 // This doesn't change because the baseFeePerGas of block where we
                 // run the eth_call is 0
                 assert.equal(
-                  rpcDataToBN(balanceResult).toString(),
+                  rpcDataToBigInt(balanceResult).toString(),
                   ethBalance.toString()
                 );
               });
@@ -670,8 +668,9 @@ contract C {
                 ]);
 
                 // The miner will get the priority fee
-                assert.isTrue(
-                  rpcDataToBN(balanceResult).eq(ethBalance.subn(500_000 * 3))
+                assert.equal(
+                  rpcDataToBigInt(balanceResult),
+                  ethBalance - 3n * 500_000n
                 );
               });
 
@@ -687,8 +686,9 @@ contract C {
                 ]);
 
                 // The miner will get the gasPrice * gas as a normalized priority fee
-                assert.isTrue(
-                  rpcDataToBN(balanceResult).eq(ethBalance.subn(500_000 * 6))
+                assert.equal(
+                  rpcDataToBigInt(balanceResult),
+                  ethBalance - 6n * 500_000n
                 );
               });
             });