+++ packages/hardhat-core/test/internal/hardhat-network/stack-traces/test.ts
@@ -1,6 +1,6 @@
-import VM from "@ethereumjs/vm";
+import { toBuffer } from "@nomicfoundation/ethereumjs-util";
+import { VM } from "@nomicfoundation/ethereumjs-vm";
 import { assert } from "chai";
-import { BN, toBuffer } from "ethereumjs-util";
 import fs from "fs";
 import path from "path";
 import semver from "semver";
@@ -26,10 +26,7 @@ import {
   SolidityStackTraceEntry,
   StackTraceEntryType,
 } from "../../../../src/internal/hardhat-network/stack-traces/solidity-stack-trace";
-import {
-  SolidityTracer,
-  SUPPORTED_SOLIDITY_VERSION_RANGE,
-} from "../../../../src/internal/hardhat-network/stack-traces/solidityTracer";
+import { SolidityTracer } from "../../../../src/internal/hardhat-network/stack-traces/solidityTracer";
 import { VmTraceDecoder } from "../../../../src/internal/hardhat-network/stack-traces/vm-trace-decoder";
 import {
   CompilerInput,
@@ -38,6 +35,7 @@ import {
 } from "../../../../src/types";
 import { setCWD } from "../helpers/cwd";
 
+import { SUPPORTED_SOLIDITY_VERSION_RANGE } from "../../../../src/internal/hardhat-network/stack-traces/constants";
 import {
   compileFiles,
   COMPILER_DOWNLOAD_TIMEOUT,
@@ -306,10 +304,10 @@ function compareStackTraces(
         `Stack trace of tx ${txIndex} entry ${i} should have value`
       );
 
-      const expectedValue = new BN(expected.value);
+      const expectedValue = BigInt(expected.value);
 
       assert.isTrue(
-        expectedValue.eq((actual as any).value),
+        expectedValue === (actual as any).value,
         `Stack trace of tx ${txIndex} entry ${i} has value ${actualValue.toString(
           10
         )} and should have ${expectedValue.toString(10)}`
@@ -329,7 +327,7 @@ function compareStackTraces(
         `Stack trace of tx ${txIndex} entry ${i} should have an errorCode`
       );
 
-      const actualErrorCodeHex = actualErrorCode.toString("hex");
+      const actualErrorCodeHex = actualErrorCode.toString(16);
 
       assert.isTrue(
         expected.errorCode === actualErrorCodeHex,