+++ packages/hardhat-core/test/internal/util/wei-values.ts
@@ -1,11 +1,10 @@
 import { assert } from "chai";
-import { BN } from "ethereumjs-util";
 
 import { weiToHumanReadableString } from "../../../src/internal/util/wei-values";
 
 describe("Wei values formatting", function () {
-  const ONE_GWEI = new BN(10).pow(new BN(9));
-  const ONE_ETH = new BN(10).pow(new BN(18));
+  const ONE_GWEI = 10n ** 9n;
+  const ONE_ETH = 10n ** 18n;
 
   it("Should show 0 wei as 0 ETH", function () {
     assert.equal(weiToHumanReadableString(0), "0 ETH");
@@ -28,11 +27,11 @@ describe("Wei values formatting", function () {
   });
 
   it("Should show 0.0001 gwei as gwei", function () {
-    assert.equal(weiToHumanReadableString(ONE_GWEI.divn(10000)), "0.0001 gwei");
+    assert.equal(weiToHumanReadableString(ONE_GWEI / 10000n), "0.0001 gwei");
   });
 
   it("Should show 0.1 gwei as gwei", function () {
-    assert.equal(weiToHumanReadableString(ONE_GWEI.divn(10)), "0.1 gwei");
+    assert.equal(weiToHumanReadableString(ONE_GWEI / 10n), "0.1 gwei");
   });
 
   it("Should show 1 gwei as gwei", function () {
@@ -40,27 +39,27 @@ describe("Wei values formatting", function () {
   });
 
   it("Should show 10 gwei as gwei", function () {
-    assert.equal(weiToHumanReadableString(ONE_GWEI.muln(10)), "10 gwei");
+    assert.equal(weiToHumanReadableString(10n * ONE_GWEI), "10 gwei");
   });
 
   it("Should show 10 gwei as gwei", function () {
-    assert.equal(weiToHumanReadableString(ONE_GWEI.muln(10)), "10 gwei");
+    assert.equal(weiToHumanReadableString(10n * ONE_GWEI), "10 gwei");
   });
 
   it("Should show 10000 gwei as gwei", function () {
-    assert.equal(weiToHumanReadableString(ONE_GWEI.muln(10000)), "10000 gwei");
+    assert.equal(weiToHumanReadableString(10_000n * ONE_GWEI), "10000 gwei");
   });
 
   it("Should show 100000 gwei as ETH", function () {
-    assert.equal(weiToHumanReadableString(ONE_GWEI.muln(100000)), "0.0001 ETH");
+    assert.equal(weiToHumanReadableString(100_000n * ONE_GWEI), "0.0001 ETH");
   });
 
   it("Should show 0.0001 ETH as ETH", function () {
-    assert.equal(weiToHumanReadableString(ONE_ETH.divn(10000)), "0.0001 ETH");
+    assert.equal(weiToHumanReadableString(ONE_ETH / 10000n), "0.0001 ETH");
   });
 
   it("Should show 0.1 ETH as ETH", function () {
-    assert.equal(weiToHumanReadableString(ONE_ETH.divn(10)), "0.1 ETH");
+    assert.equal(weiToHumanReadableString(ONE_ETH / 10n), "0.1 ETH");
   });
 
   it("Should show 1 ETH as ETH", function () {
@@ -69,12 +68,12 @@ describe("Wei values formatting", function () {
 
   it("Should show 1.2 ETH as ETH", function () {
     assert.equal(
-      weiToHumanReadableString(ONE_ETH.add(ONE_ETH.divn(10).muln(2))),
+      weiToHumanReadableString(ONE_ETH + (2n * ONE_ETH) / 10n),
       "1.2 ETH"
     );
   });
 
   it("Should show 43 ETH as ETH", function () {
-    assert.equal(weiToHumanReadableString(ONE_ETH.muln(43)), "43 ETH");
+    assert.equal(weiToHumanReadableString(43n * ONE_ETH), "43 ETH");
   });
 });