+++ packages/hardhat-network-helpers/CHANGELOG.md
@@ -1,5 +1,17 @@
 # @nomicfoundation/hardhat-network-helpers
 
+## 1.0.6
+
+### Patch Changes
+
+- 89f153a72: Fix `setStorageAt` so it can accept multiple leading zeros in the slot
+
+## 1.0.5
+
+### Patch Changes
+
+- f3ba15ca8: Added a new `setPrevRandao` helper
+
 ## 1.0.4
 
 ### Patch Changes